KAST_RKN

+++ models/KAST.py
@@ -135,7 +135,7 @@ class KAST(tf.keras.Model):
         mask = np.random.binomial(1, 0.9, [bs, seq_size])
         mask[:, 0] = 1
 
-        mask = tf.reshape(mask, [bs, seq_size, 1, 1, 1])
+        mask = tf.cast(tf.reshape(mask, [bs, seq_size, 1, 1, 1]), tf.float32)
 
         k = k*mask
